login completo e ajuste comanda serviço

+++ src/service/AvaliacoesService.ts
@@ -1,10 +1,11 @@
 import {  typeAvaliacao, typeQuestions } from '../components/Interface'
 import { api } from './Api'
+import { getToken } from './authService'
 export async function getAvaliacoes() {
-    return await api.get<[]>('/avaliacoes')
+    return await api.get<[]>('/api/avaliacoes',{headers:{"Authorization": `Bearer ${getToken()}`}})
 }
 export async function postAvaliacoes(avaliacoes: typeAvaliacao) {
-    const resp = await api.post('/avaliacoes', avaliacoes)
+    const resp = await api.post('/api/avaliacoes', avaliacoes,{headers:{"Authorization": `Bearer ${getToken()}`}})
     return resp.status === 201
 }
 type typeAvaliacoesQuestion={
@@ -12,19 +13,19 @@ type typeAvaliacoesQuestion={
     questions:typeQuestions[]
 }
 export async function getOneAvaliacoesQuestion(id_avaliacoes: number) {
-    return await api.get<typeAvaliacoesQuestion>(`/avaliacoes/questions/${id_avaliacoes}`)
+    return await api.get<typeAvaliacoesQuestion>(`/api/avaliacoes/questions/${id_avaliacoes}`,{headers:{"Authorization": `Bearer ${getToken()}`}})
 }
 export async function getAvaliacoesItenQuestions(id_avaliacoes: number) {
-    return await api.get<typeQuestions[]>(`/avaliacoes/itensquestions/${id_avaliacoes}`)
+    return await api.get<typeQuestions[]>(`/api/avaliacoes/itensquestions/${id_avaliacoes}`,{headers:{"Authorization": `Bearer ${getToken()}`}})
 }
 
 export async function putAvaliacoes(id_avaliacoes: number, avaliacoes: typeAvaliacao) {
-    const resp = await api.put<typeAvaliacao>(`/avaliacoes/${id_avaliacoes}`, avaliacoes)
+    const resp = await api.put<typeAvaliacao>(`/api/avaliacoes/${id_avaliacoes}`, avaliacoes,{headers:{"Authorization": `Bearer ${getToken()}`}})
     return resp.status === 204
 }
 export async function deleteAvaliacoes(id_avaliacoes: number) {
     if (window.confirm('Deseja Realmente excluir Avaliacao?')) {
-        const resp = (await api.delete(`/avaliacoes/${id_avaliacoes}`))
+        const resp = (await api.delete(`/api/avaliacoes/${id_avaliacoes}`,{headers:{"Authorization": `Bearer ${getToken()}`}}))
         return resp.status === 204
     }
     return
